requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
aiohttp>=3.8.4
google-generativeai>=0.3.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=1.10.11
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
psutil>=5.9.0
